+++ utilities/src/models/pixelAuto.ts
@@ -1,4 +1,4 @@
-import type { Pixel } from "@/utilities/types/pixel";
+import type { Pixel } from "@/utilities/src/types/pixel";
 
 export class AutoPixel implements Pixel {
   private readonly _value = "auto";
