+++ utilities/src/types/validator.ts
