+++ utilities/tests/models/pixelAuto.test.ts
@@ -1,6 +1,6 @@
 import { test } from "vitest";
 
-import { AutoPixel } from "@/utilities/models/pixelAuto";
+import { AutoPixel } from "@/utilities/src/models/pixelAuto";
 
 test("AutoPixel should create an instance with value 'auto'", ({ expect }) => {
   const testee = AutoPixel.create();
